requirements.txt
strawberry-graphql==0.246.2
fastapi==0.115.2